fix: Added isOrganizerAdmin

+++ src/services/actions/management-tool/channels/twitter/deleteTwitterPost.ts
@@ -1,8 +1,15 @@
 'use server'
 
+import { isOrganizerAdmin } from '@/lib/auth/lucia'
 import prisma from '@/lib/prisma'
 
-export const deleteTwitterPost = async (postId: number) => {
+export const deleteTwitterPost = async (postId: number, userId: string, organizationId: string) => {
+
+  const isAllowed = await isOrganizerAdmin(organizationId, userId)
+  if (!isAllowed) {
+    throw new Error('Unauthorized: You don´t have permission.')
+  }
+
   try {
     const deletedPost = await prisma.post.delete({
       where: { id: postId }
